Added a LOT of debug code

+++ noextern.go
@@ -33,10 +33,10 @@ func (m *ExternMetadata) DoWork() error {
 
 // Get gets a previously allocated memory slab of the provided size. If no memories of that size exist,
 // it returns a NoOpError. The caller is then responsible for allocating the memory themselves.
-func (m *ExternMetadata) Get(dev Device, size uint) (Memory, error) { return nil, noopError{} }
+func (m *ExternMetadata) Get(dev Device, size int64) (Memory, error) { return nil, noopError{} }
 
 // Put puts a previously allocated memory slab of the provided size back into the pool. Currently this is a No-op in this build.
-func (m *ExternMetadata) Put(dev Device, mem Memory, size uint) {}
+func (m *ExternMetadata) Put(dev Device, mem Memory, size int64) {}
 
 // Cleanup cleans up the ancillary allocations made during the calling of batched external device function.
 //
